Clean up admin pages
