+++ src/routes/(auth)/admin/actions/+page.server.ts
@@ -1,5 +1,6 @@
-import { error } from '@sveltejs/kit';
+import { error, fail, type Actions } from '@sveltejs/kit';
 import type { PageServerLoad } from './$types';
+import { DELETE } from '$lib/api/elite';
 
 export const load = (async ({ parent, locals }) => {
 	const { user, session } = await parent();
@@ -13,3 +14,25 @@ export const load = (async ({ parent, locals }) => {
 		user,
 	};
 }) satisfies PageServerLoad;
+
+export const actions: Actions = {
+	clearcontests: async ({ locals }) => {
+		const { access_token: token } = locals;
+
+		if (!token) {
+			throw error(404, 'Not Found');
+		}
+
+		const { response } = await DELETE(`/admin/upcomingcontests`, {
+			headers: { Authorization: `Bearer ${token}` },
+		});
+
+		if (!response.ok) {
+			return fail(500, { error: 'Failed to delete contests' });
+		}
+
+		return {
+			success: true,
+		};
+	},
+};
